# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Xlarge Api",
        "description": "xlarge laravel api Documentation",
        "contact": {
            "name": "github",
            "url": "https://github.com/menaemg/xlarge"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/user/comments": {
            "get": {
                "tags": [
                    "user comments"
                ],
                "summary": "get all user comments",
                "description": "get all user comments",
                "operationId": "user.comments",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/user/replies": {
            "get": {
                "tags": [
                    "user replies"
                ],
                "summary": "get all user replies",
                "description": "get all user replies",
                "operationId": "user.replies",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/editor/comments": {
            "get": {
                "tags": [
                    "editor comments"
                ],
                "summary": "get all comments",
                "description": "get all comments",
                "operationId": "editor.comments",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/editor/replies": {
            "get": {
                "tags": [
                    "editor replies"
                ],
                "summary": "get all replies",
                "description": "get all replies",
                "operationId": "editor.replies",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}